+++ examples/logs/modules.tf
@@ -1,4 +1,4 @@
-module "azure-region" {
+module "azure_region" {
   source  = "claranet/regions/azurerm"
   version = "2.0.1"
 
@@ -9,7 +9,7 @@ module "rg" {
   source  = "claranet/rg/azurerm"
   version = "2.1.0"
 
-  location    = module.azure-region.location
+  location    = module.azure_region.location
   client_name = var.client_name
   environment = var.environment
   stack       = var.stack
@@ -20,8 +20,8 @@ module "logs" {
   version = "2.0.3"
 
   client_name    = var.client_name
-  location       = module.azure-region.location
-  location_short = module.azure-region.location_short
+  location       = module.azure_region.location
+  location_short = module.azure_region.location_short
   environment    = var.environment
   stack          = var.stack